update: separate commands
